refactor(engine): extract UsesEnvConfig to shared engine package
